auth-service/requirements.txt
fastapi==0.114.0
uvicorn==0.35.0
boto3==1.28.61
botocore==1.31.61
python-jose==3.3.0
requests==2.32.0
